# docs/swagger.yaml
basePath: /
definitions:
  github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_rede_world-cup-quiniela_internal_service.DLQEntry:
    properties:
      dead_lettered_at:
        type: string
      handler_err:
        type: string
      payload:
        additionalProperties: {}
        type: object
    type: object
  github_com_rede_world-cup-quiniela_internal_service.DLQStat:
    properties:
      count:
        type: integer
      event_type:
        type: string
      oldest_at:
        type: string
      sample:
        items:
          $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_service.DLQEntry'
        type: array
    type: object
...
  internal_api_handler.AuditLogResponse:
    properties:
      action:
        type: string
      actor_id:
        type: integer
      actor_role:
        type: string
      created_at:
        type: string
      id:
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      resource_id:
        type: integer
      resource_type:
        type: string
    type: object
...
  internal_api_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail'
    type: object
...
  internal_api_handler.PageMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  internal_api_handler.Paged-internal_api_handler_AuditLogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_api_handler.AuditLogResponse'
        type: array
      page:
        $ref: '#/definitions/internal_api_handler.PageMeta'
    type: object
...
  internal_api_handler.replayRequest:
    properties:
      limit:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the World Cup prediction game. Manage fixtures, submit
    score forecasts, and track leaderboards.
  title: World Cup Quiniela API
  version: "1.0"
paths:
  /api/v1/admin/audit-log:
    get:
      description: Returns a paginated list of audit log entries. Supports filtering
      parameters:
      - description: Filter by actor (admin) user ID
        in: query
        name: actor_id
        type: integer
      - description: Filter by action string (e.g. admin_user.banned)
        in: query
        name: action
        type: string
      - description: Filter by resource type (e.g. membership, payment_record)
        in: query
        name: resource_type
        type: string
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: integer
      - description: Max records per page (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.Paged-internal_api_handler_AuditLogResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit log entries
      tags:
      - admin-audit-log
  /api/v1/admin/audit-log/entity/{type}/{id}:
    get:
      description: Returns paginated audit log entries scoped to a single resource
      parameters:
      - description: Resource type (e.g. membership, payment_record, user)
        in: path
        name: type
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Max records per page (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.Paged-internal_api_handler_AuditLogResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid resource ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit log by entity
      tags:
      - admin-audit-log
...
  /api/v1/admin/dlq:
    delete:
      description: Permanently deletes all entries from all DLQ keys. Use with
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              removed:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Purge dead-letter queue
      tags:
      - admin-dlq
    get:
      description: Returns the count, oldest entry age, and a sample of messages for
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_service.DLQStat'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dead-letter queue stats
      tags:
      - admin-dlq
  /api/v1/admin/dlq/replay:
    post:
      consumes:
      - application/json
      description: Re-enqueues up to limit entries from all DLQ keys back onto their
      parameters:
      - description: Optional replay limit (default 10)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_api_handler.replayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              replayed:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replay dead-letter queue
      tags:
      - admin-dlq
...
securityDefinitions:
  BearerAuth:
    description: 'Clerk JWT token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"